finish preview pdf

+++ index.php
@@ -1,4 +1,9 @@
 <?php
+// Ajouter ce code au début de votre fichier
+error_reporting(E_ALL);
+ini_set('display_errors', '1');
+
+
 require_once "./controllers/UserController.php";
 $uri = parse_url($_SERVER["REQUEST_URI"], PHP_URL_PATH);
 
@@ -6,7 +11,7 @@ if ($uri == "/" || $uri == "/index.php" || $uri == "/list") {
     header('Content-Type: text/html; charset=utf-8');
 
     if (isset($_POST)) {
-        if ($_POST['deleted_ids']) {
+        if (isset($_POST['deleted_ids']) && $_POST['deleted_ids']) {
             delete_users($_POST["deleted_ids"]);
         }
         if (isset($_POST["delete_user_id"])) {
@@ -34,7 +39,7 @@ if ($uri == "/create") {
     header('Content-Type: text/html; charset=utf-8');
 
     if (isset($_POST)) {
-        if ($_POST["action"] == "create") {
+        if (isset($_POST["action"]) &&  $_POST["action"] == "create") {
             show_msg_user_created($_POST);
             // Redirection apres 5
             header("Refresh:5; url=/list");
@@ -46,7 +51,7 @@ if ($uri == "/create") {
 if ($uri == "/editUser/") {
 
     if (isset($_POST)) {
-        if (preg_match_all("/edit/i", $_POST["action"]) && $_POST["id"]) {
+        if (isset($_POST["action"]) && preg_match_all("/edit/i", $_POST["action"]) && $_POST["id"]) {
             show_msg_update_sucessfully($_POST);
             // Redirection apres 5
             header("Refresh:5; url=/details/?id=" . $_POST['id']);
@@ -64,4 +69,11 @@ if ($uri == "/dist/") {
     header('Content-Type: application/javascript; charset=utf-8');
     readfile('dist/bundle.js');
     exit;
+}
+
+if ($uri == "/pdf_list") {
+    header("Content-Type: application/pdf");
+    require_once "./fpdf/fpdf.php";
+    show_pdf_list();
+    var_dump(show_pdf_list());
 }
